added thymeleaf to the backend

+++ src/main/java/com/thymeleaf/prueba/entities/Student.java
@@ -23,11 +23,11 @@ public class Student{
         this.email = email;
     }
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Integer id) {
         this.id = id;
     }
